add Ubuntu2404 support

+++ scripts/cleanup-managedimages.ps1
@@ -8,7 +8,8 @@ param(
     [String] [Parameter (Mandatory=$true)] $VmssNameWindows,
     [String] [Parameter (Mandatory=$true)] $VmssNameWindows2022,
     [String] [Parameter (Mandatory=$true)] $VmssNameUbuntu,
-    [String] [Parameter (Mandatory=$true)] $VmssNameUbuntu2204
+    [String] [Parameter (Mandatory=$true)] $VmssNameUbuntu2204,
+    [String] [Parameter (Mandatory=$true)] $VmssNameUbuntu2404
 )
 
 az login --service-principal --username $ClientId --password $ClientSecret --tenant $TenantId | Out-Null
@@ -20,10 +21,11 @@ $windowsManagedImage = az vmss show --name $VmssNameWindows --resource-group $Ag
 $windowsManagedImage2022 = az vmss show --name $VmssNameWindows2022 --resource-group $AgentsResourceGroup --query 'virtualMachineProfile.storageProfile.imageReference.id' --subscription $SubscriptionId | Out-String | ConvertFrom-Json
 $ubuntuManagedImage = az vmss show --name $VmssNameUbuntu --resource-group $AgentsResourceGroup --query 'virtualMachineProfile.storageProfile.imageReference.id' --subscription $SubscriptionId | Out-String | ConvertFrom-Json
 $ubuntuManagedImage2204 = az vmss show --name $VmssNameUbuntu2204 --resource-group $AgentsResourceGroup --query 'virtualMachineProfile.storageProfile.imageReference.id' --subscription $SubscriptionId | Out-String | ConvertFrom-Json
+$ubuntuManagedImage2404 = az vmss show --name $VmssNameUbuntu2404 --resource-group $AgentsResourceGroup --query 'virtualMachineProfile.storageProfile.imageReference.id' --subscription $SubscriptionId | Out-String | ConvertFrom-Json
 
 
 foreach ($managedImage in $managedImages) {
-    if ($managedImage -ne $windowsManagedImage -and $managedImage -ne $ubuntuManagedImage -and $managedImage -ne $windowsManagedImage2022 -and $managedImage -ne $ubuntuManagedImage2204) {
+    if ($managedImage -ne $windowsManagedImage -and $managedImage -ne $ubuntuManagedImage -and $managedImage -ne $windowsManagedImage2022 -and $managedImage -ne $ubuntuManagedImage2204 -and $managedImage -ne $ubuntuManagedImage2404) {
         Write-Host "Found a match, deleting orphaned managed image: $managedImage"
         az image delete --ids $managedImage | Out-Null
     }
